Create a proper cache facility (renaming utils module to cache in the process).
  * trunk/src/shakespeare/cache.py, trunk/src/shakespeare/utils.py, trunk/src/shakespeare/utils_test.py, trunk/src/shakespeare/cache_test.py: rename utils to cache and refactor to have a proper Cache class
  * trunk/src/shakespeare/dm.py:
    * class Material: change cache_path attribute to url

All other changes are just knock-on renames/rewritings stemming from those just mentioned:

  * trunk/src/shakespeare/cherrypy_handler.py
  * trunk/src/shakespeare/textutils.py
  * trunk/src/shakespeare/moby.py
  * trunk/src/shakespeare/concordance.py
  * trunk/src/shakespeare/gutenberg_test.py
  * trunk/src/shakespeare/gutenberg.py

One side effect of the use of change to url from cache_path is that we now see the original (uncleaned) gutenberg texts. This will be fixed in a subsequent commit (as it may involve some substantial changes).